Refactor JSV behat structure to use behat.local.yml setup
