+++ includes/bootstrap/SWSFeatureContext.php
@@ -363,4 +363,19 @@ class SWSFeatureContext extends Drupal\DrupalExtension\Context\DrupalContext {
     $ctx->fillField($radioId, $radioButton->getAttribute('value'));
   }
 
+  /**
+   * Find an element in a region.
+   * see http://cgit.drupalcode.org/panopoly/tree/tests/behat/features/bootstrap/FeatureContext.php?id=18a2ccbdad8c8064aa36f8c57ae7416ee018b92f
+   *
+   * @Then /^I should see a "([^"]*)" element in the "([^"]*)" region$/
+   */
+  public function assertRegionElement($tag, $region) {
+    $regionObj = $this->getRegion($region);
+    $elements = $regionObj->findAll('css', $tag);
+    if (!empty($elements)) {
+      return;
+    }
+    throw new \Exception(sprintf('The element "%s" was not found in the "%s" region on the page %s', $tag, $region, $this->getSession()->getCurrentUrl()));
+  }
+
 }
